fully remove django from ubersystem core - in order to do this, needed to modify the way we were ingesting config data, which required moving a few things around - moved all jinja init to jinja.py - cleaned up config settings, removed django
