+++ requirements.txt
@@ -1,4 +1,3 @@
-Django==1.6.1
 psycopg2==2.5.2
 py3k-bcrypt==0.3
 stripe==1.11.0
