+++ uber/configspec.ini
@@ -7,6 +7,11 @@
 # after a deadline or on-site.
 dev_box = boolean(default=True)
 
+# List of directories we should look in for our templates
+# This list is appeneded to by plugins, we'll look for a template
+# in the following directories, stopping once one is found
+template_dirs = string_list(default=list('%(module_root)s/templates'))
+
 # This turns on our automated emails.  See the description in the [secret]
 # section below for an explanation of how this works.
 send_emails = boolean(default=False)
@@ -598,13 +603,6 @@ path = string(default="")
 3 = string(default="signups/food_item.html")
 4 = string(default="signups/shifts_item.html")
 __many__ = string(default="")
-
-
-[django]
-# These settings are passed to django.settings.configure()
-DEBUG = boolean(default=True)
-TEMPLATE_DEBUG = boolean(default=True)
-TEMPLATE_DIRS = string_list(default=list('%(module_root)s/templates'))
 
 
 [appconf]
